Extract document/item-sheet classes into src/ modules
Move TheFadeActor, TheFadeItem, and TheFadeItemSheet out of thefade.js
into dedicated files. No logic change — classes are relocated as-is and
re-imported by the entry file.

- src/actor.js (TheFadeActor) — imports BODY_PARTS, FALLBACK_ACTOR_DATA
- src/item.js (TheFadeItem) — imports DEFAULT_WEAPON, DEFAULT_ARMOR, DEFAULT_SKILL
- src/item-sheet.js (TheFadeItemSheet) — imports SIZE_OPTIONS,
  PATH_SKILL_TYPES, DEFAULT_SKILLS from constants; getRankValue and
  openCompendiumBrowser from helpers.js

thefade.js: 7934 → 5352 lines. What remains is TheFadeCharacterSheet,
registerItemSheets, and the system hooks — a more manageable entry
point. The character-sheet god-class split is the next step.
